check q30

+++ modules/local/get_raw_stats.nf
@@ -20,23 +20,22 @@ process GET_RAW_STATS {
     def num2 = "${reads[1]}".minus(".fastq.gz")
 
     """
-    echo "FAIL" > ${prefix}_result.txt
 
     q30.py ${reads[0]} > ${prefix}_R1_stats.txt
-    
     q30.py ${reads[1]} > ${prefix}_R2_stats.txt
 
-    mv ${reads[0]} ${num1}C.fastq.gz
-    mv ${reads[1]} ${num2}C.fastq.gz
-
-    create_raw_stats_output.py -n ${prefix} -r1 ${prefix}_R1_stats.txt -r2 ${prefix}_R2_stats.txt
     
-    comb_stats_chk.py -r ${prefix}_raw_read_counts.txt
+    if [ -f ${prefix}_R2_stats.txt -a -f ${prefix}_R1_stats.txt ];
+    then   
+        create_raw_stats_output.py -n ${prefix} -r1 ${prefix}_R1_stats.txt -r2 ${prefix}_R2_stats.txt
+        comb_stats_chk.py -r ${prefix}_raw_read_counts.txt
+
+        mv ${reads[0]} ${num1}C.fastq.gz
+        mv ${reads[1]} ${num2}C.fastq.gz
     
-    mv ${num1}C.fastq.gz ${num1}.fastq.gz
-    mv ${num2}C.fastq.gz ${num2}.fastq.gz
-
-
+        mv ${num1}C.fastq.gz ${num1}.fastq.gz
+        mv ${num2}C.fastq.gz ${num2}.fastq.gz
+    fi
 
     cat <<-END_VERSIONS > versions.yml
     "${task.process}":
